refactor(documents): Split desktop and mobile views into separate components (DocumentsDesktop, DocumentsMobile). Remove redundant handler redefinitions.
